Move web share shim under webCompat
